+++ test/open_route_service_action_tests.dart
@@ -2,4 +2,7 @@ import 'package:test/test.dart';
 
 import 'open_route_service_test.dart' show apiKey;
 
-void main() => test('Validate API Key', () async => expect(apiKey, 'test'));
+void main() => test(
+      'Validate API Key Reset',
+      () async => expect(apiKey, 'test'),
+    );